# Improve Gauge and Counter performance.