@beorn7 2015-02-02
![benchcmp](https://cloud.githubusercontent.com/assets/5609886/6004978/29d58240-ab07-11e4-81a6-b0d9598a0b66.png)