@juliusv 2015-02-02
Currently at the Imbiss, but just needed to say: Whoa. That's frigging
awesome!
On Feb 2, 2015 6:13 PM, "Björn Rabenstein" <notifications@github.com> wrote:

> [image: benchcmp]
> <https://cloud.githubusercontent.com/assets/5609886/6004978/29d58240-ab07-11e4-81a6-b0d9598a0b66.png>
>
> —
> Reply to this email directly or view it on GitHub
> <https://github.com/prometheus/client_golang/pull/64#issuecomment-72495859>
> .
>