@brian-brazil 2015-02-02
:+1: 

Using labels is still slow, can the user cache the result of the wWith()?