@beorn7 2015-02-02
@brian-brazil Yes, the user can just keep the metric retrieved with With(). That's also documented somewhere.