@brian-brazil 2015-02-02
It'd be good to benchmark that too, so users can make the complexity vs. cpu tradeoff