@beorn7 2015-02-02
@brian-brazil Once you have the metric retrieved with With(), there is no difference to a label-less metric. It's the same code.